fix: [IABT-1452] Test account login loop (#4512)
This PR fixes a bug which occurs when selecting an actual IDP or CIE
login before trying to login with the test account. In this case, the
login would fail and the user would be redirected to the methods
selection screen.

- ts/features/lollipop/saga/index.ts: Checks whether or not the user is
logged in with a test IDP;
- ts/sagas/testLoginSaga.ts: Dispatches an action with a test IDP when
logging with a test account;
- ts/screens/authentication/IdpSelectionScreen.tsx: Sets the selected
IDP to the test one when trying to login with a test account;
- ts/store/reducers/authentication.ts: Edits the selector in order to
check if the user is authenticated with the test IDP;
- ts/features/lollipop/__tests__/login.test.ts: Updates the test case
accordingly.
- ts/models/IdentityProvider.ts removed in favour of SpidIdp
(definitions/content/SpidIdp.ts) and IdpData
(definitions/content/IdpData.ts)

- Login into the test account by tapping 5 times on the `Select your
identity provider` header;
- Logout. Select an actual IDP, then navigate back and login again in
the test account;
- Logout. Select `Login with CIE`, then navigate back and login again
with the test account;
- Logout. Try to login normally with both CIE and SPID.

---------

Co-authored-by: Andrea Piai <andrea.piai@pagopa.it>
Co-authored-by: LazyAfternoons <LazyAfternoons@users.noreply.github.com>
Co-authored-by: Cristiano Tofani <cri.tofani@gmail.com>